pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "genai-platform"
version = "0.1.0"
description = "A platform for building production-ready GenAI applications"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.setuptools]
packages = {find = {}}

[tool.setuptools.package-data]
"proto" = ["*.proto"]

